+++ semantic-conventions/src/opentelemetry/semconv/templating/markdown/__init__.py
@@ -39,6 +39,10 @@ from opentelemetry.semconv.model.semantic_convention import (
 from opentelemetry.semconv.model.utils import ID_RE
 from opentelemetry.semconv.templating.markdown.options import MarkdownOptions
 
+_REQUIREMENT_LEVEL_URL = (
+    "https://opentelemetry.io/docs/specs/semconv/general/attribute-requirement-level/"
+)
+
 
 class RenderContext:
     def __init__(self):
@@ -80,10 +84,6 @@ class MarkdownRenderer:
     ]
 
     prelude = "<!-- semconv {} -->\n"
-    table_headers = "| Attribute  | Type | Description  | Examples  | Requirement Level |\n|---|---|---|---|---|\n"
-    table_headers_omitting_req_level = (
-        "| Attribute  | Type | Description  | Examples  |\n|---|---|---|---|\n"
-    )
 
     def __init__(
         self, md_folder, semconvset: SemanticConventionSet, options=MarkdownOptions()
@@ -99,6 +99,16 @@ class MarkdownRenderer:
         # We build the dict that maps each attribute that has to be rendered to the latest visited file
         # that contains it
         self.filename_for_attr_fqn = self._create_attribute_location_dict()
+
+        req_level = f"[Requirement Level]({_REQUIREMENT_LEVEL_URL})"
+
+        self.table_headers = (
+            f"| Attribute  | Type | Description  | Examples  | {req_level} |"
+            "\n|---|---|---|---|---|\n"
+        )
+        self.table_headers_omitting_req_level = (
+            "| Attribute  | Type | Description  | Examples  |\n|---|---|---|---|\n"
+        )
 
     def to_markdown_attr(
         self,
@@ -169,14 +179,14 @@ class MarkdownRenderer:
 
     def derive_requirement_level(self, attribute: SemanticAttribute):
         if attribute.requirement_level == RequirementLevel.REQUIRED:
-            required = "Required"
+            required = "`Required`"
         elif attribute.requirement_level == RequirementLevel.CONDITIONALLY_REQUIRED:
             if len(attribute.requirement_level_msg) < self.options.break_count:
-                required = "Conditionally Required: " + attribute.requirement_level_msg
+                required = "`Conditionally Required` " + attribute.requirement_level_msg
             else:
                 # We put the condition in the notes after the table
                 self.render_ctx.add_note(attribute.requirement_level_msg)
-                required = f"Conditionally Required: [{len(self.render_ctx.notes)}]"
+                required = f"`Conditionally Required` [{len(self.render_ctx.notes)}]"
         elif attribute.requirement_level == RequirementLevel.OPT_IN:
             required = "Opt-In"
         else:  # attribute.requirement_level == Required.RECOMMENDED or None
@@ -188,20 +198,20 @@ class MarkdownRenderer:
                 required = "See below"
             else:
                 if not attribute.requirement_level_msg:
-                    required = "Recommended"
+                    required = "`Recommended`"
                 elif len(attribute.requirement_level_msg) < self.options.break_count:
-                    required = "Recommended: " + attribute.requirement_level_msg
+                    required = "`Recommended` " + attribute.requirement_level_msg
                 else:
                     # We put the condition in the notes after the table
                     self.render_ctx.add_note(attribute.requirement_level_msg)
-                    required = f"Recommended: [{len(self.render_ctx.notes)}]"
+                    required = f"`Recommended` [{len(self.render_ctx.notes)}]"
         return required
 
     def write_table_header(self, output: io.StringIO):
         if self.render_ctx.is_omit_requirement_level:
-            output.write(MarkdownRenderer.table_headers_omitting_req_level)
+            output.write(self.table_headers_omitting_req_level)
         else:
-            output.write(MarkdownRenderer.table_headers)
+            output.write(self.table_headers)
 
     def to_markdown_attribute_table(
         self, semconv: BaseSemanticConvention, output: io.StringIO
@@ -322,8 +332,8 @@ class MarkdownRenderer:
             if enum.custom_values:
                 output.write(
                     "has the following list of well-known values."
-                    + " If one of them applies, then the respective value MUST be used,"
-                    + " otherwise a custom value MAY be used."
+                    + " If one of them applies, then the respective value MUST be used;"
+                    + " otherwise, a custom value MAY be used."
                 )
             else:
                 output.write("MUST be one of the following:")